requirements.txt
fastapi
uvicorn
starlette
pydantic
vaderSentiment
textstat
fastmcp
openai
gradio
python-multipart
google-generativeai
openenv-core
python-dotenv